Relicensed LGPL.
	* gladeui/glade-name-context.[ch], gladeui/glade-editable.[ch]: Relicensed LGPL.

	* plugins/gtk+/glade-image-editor.c, plugins/gtk+/glade-image-item-editor.c,
	plugins/gtk+/glade-button-editor.c, plugins/gtk+/glade-label-editor.c,
	plugins/gtk+/glade-cell-renderer-editor.c, plugins/gtk+/glade-tool-button-editor.c,
	gladeui/glade-editor-table.c, gladeui/glade-base-editor.c: 
	Now editable portions of properties dont expand in tables, leaving more freedom 
	for labels to expand in tight spots, no more item_labels in radio buttons (they dont expand).

	* gladeui/glade-widget-adaptor.c: Fixed regression, now disabled classes are not
	reinstalled in subclasses.


svn path=/trunk/; revision=2065

+++ gladeui/glade-project.c
@@ -167,8 +167,8 @@ typedef struct {
 
 static void         glade_project_set_target_version       (GladeProject *project,
 							    const gchar  *catalog,
-							    gint          major,
-							    gint          minor);
+							    guint16       major,
+							    guint16       minor);
 static void         glade_project_get_target_version       (GladeProject *project,
 							    const gchar  *catalog,
 							    gint         *major,
@@ -912,10 +912,10 @@ glade_project_fix_object_props (GladeProject *project)
 
 static gchar *
 glade_project_read_requires_from_comment (GladeXmlNode  *comment,
-					  gint          *major,
-					  gint          *minor)
+					  guint16       *major,
+					  guint16       *minor)
 {
-	gint maj, min;
+	gint   maj, min;
 	gchar *value, buffer[256];
 	gchar *required_lib = NULL;
 
@@ -949,7 +949,7 @@ glade_project_read_requires (GladeProject *project,
 	GladeXmlNode *node;
 	gchar        *required_lib = NULL;
 	gboolean      loadable = TRUE;
-	gint          major, minor;
+	guint16       major, minor;
 
 	for (node = glade_xml_node_get_children_with_comments (root_node); 
 	     node; node = glade_xml_node_next_with_comments (node))
@@ -2654,8 +2654,8 @@ glade_project_get_target_version (GladeProject *project,
 static void
 glade_project_set_target_version (GladeProject *project,
 				  const gchar  *catalog,
-				  gint          major,
-				  gint          minor)
+				  guint16       major,
+				  guint16       minor)
 {
 	GladeTargetableVersion *version;
 	GSList *radios, *list;
